backend/requirements.txt
asl-detection==0.1.0
flask==2.0.1
flask-cors==3.0.10
